Updates to provisioning and deprovisioning scripts and templates

+++ Learning Modules/ProvisionTenants/Demo-ProvisionTenants.ps1
@@ -4,16 +4,17 @@
 # http://events.wtp-mt.<USER>.trafficmanager.net
 
 # Parameters for scenarios #1 and #2, provision or deprovision a single tenant 
-$TenantName = "Brown Maple Racing" #  name of the venue to be added/removed as a tenant
-$VenueType = "motorracing" # valid types: blues, classicalmusic, dance, jazz, judo, motorracing, multipurpose, opera, rockmusic, soccer 
+$TenantName = "Red Maple Racing" #  name of the venue to be added/removed as a tenant
+$VenueType  = "motorracing" # valid types: blues, classicalmusic, dance, jazz, judo, motorracing, multipurpose, opera, rockmusic, soccer 
 $PostalCode = "98052"
 
 $Scenario = 1
 <# Select the scenario to run
     #    Scenario
-    1       Provision a single tenant
-    2       Remove a provisioned tenant
-    3       Provision a batch of tenants
+    1       Provision a tenant in the default multi-tenant database
+    2       Provision a tenant in a new single-tenant database
+    3       Remove a provisioned tenant
+    4       Provision a batch of tenants
 #>
 
 ## ------------------------------------------------------------------------------------------------
@@ -26,7 +27,7 @@ Import-Module "$PSScriptRoot\..\UserConfig" -Force
 # Get Azure credentials if not already logged on,  Use -Force to select a different subscription 
 Initialize-Subscription -NoEcho
 
-### Provision a single tenant
+### Provision a tenant in the default multi-tenant database
 if ($Scenario -eq 1)
 {
     New-Tenant `
@@ -45,8 +46,26 @@ if ($Scenario -eq 1)
 }
 #>
 
-### Remove a provisioned tenant
+### Provision a tenant in a single-tenant database
 if ($Scenario -eq 2)
+{
+    & $PSScriptRoot\New-TenantAndDatabase `
+        -TenantName $TenantName `
+        -VenueType $VenueType `
+        -PostalCode $PostalCode `
+        -ErrorAction Stop `
+        > $null
+
+    Write-Output "Provisioning complete for tenant '$TenantName'"
+
+    # Open the events page for the new venue
+    # Start-Process "http://events.wtp.$($wtpUser.Name).trafficmanager.net/$(Get-NormalizedTenantName $TenantName)"
+    
+    exit
+}
+
+### Remove a provisioned tenant
+if ($Scenario -eq 3)
 {
     & $PSScriptRoot\Remove-ProvisionedTenant.ps1 -TenantName $TenantName 
     exit
@@ -54,7 +73,7 @@ if ($Scenario -eq 2)
 }
 
 ### Provision a batch of tenants
-if ($Scenario -eq 3)
+if ($Scenario -eq 4)
 {
     $config = Get-Configuration
